requirements.txt
apache-airflow==2.9.2
apache-airflow-providers-postgres==5.11.2
pandas==2.2.2
requests==2.32.3
psycopg2-binary==2.9.9
apache-airflow-providers-amazon==8.21.0
apache-airflow-providers-google==10.19.0
apache-airflow-providers-microsoft-azure==8.2.0
apache-airflow-providers-sftp==4.10.1
apache-airflow-providers-http==4.12.0
apache-airflow-providers-docker==3.12.1
apache-airflow-providers-slack==8.7.1
boto3==1.34.0
paramiko==3.4.0
